backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
nltk = "^3.8.1"
numpy = "^2.0.0"
Django = "^5.0.6"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
tzdata = "^2024.1"
django-cors-headers = "^4.4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
